refactor: Split monolithic routes.ts into domain-specific modules
Split the 1,487-line routes.ts file into 9 focused, domain-specific route modules
for improved maintainability, testability, and code organization.

Changes:
- Created server/routes/ directory structure
- Extracted 49 routes into 7 domain-specific files:
  * auth-routes.ts (459 lines) - Authentication & password reset
  * forum-routes.ts (140 lines) - Forum categories, topics, posts
  * alert-routes.ts (78 lines) - Price alerts management
  * product-routes.ts (418 lines) - Product search, price history, analytics
  * retailer-routes.ts (23 lines) - Retailer data endpoints
  * admin-routes.ts (344 lines) - Admin analytics & management
  * health-routes.ts (44 lines) - Health check endpoints
- Created helpers.ts (49 lines) - Shared auth middleware (withAuth, withAdmin)
- Created index.ts (46 lines) - Route aggregation and registration
- Backed up original routes.ts to routes.ts.backup

Benefits:
✅ Improved maintainability - Each file has single responsibility
✅ Better testability - Can test route modules in isolation
✅ Easier navigation - Developers can find routes by domain
✅ Reduced merge conflicts - Team can work on different domains
✅ Better code organization - Clear separation of concerns

Total: 1,601 lines across 9 files vs 1,487 lines in one file
No functional changes - All routes work identically

Part of Phase 2: Code Quality & Architecture improvements
Addresses audit finding H6: Monolithic route file
